% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beta_95}
\alias{beta_95}
\title{Beta equation of development rate as a function of temperature.}
\format{
A list of eight elements describing the equation.
\describe{
  \item{eq}{The equation (formula object).}
  \item{eqAlt}{The equation (string).}
  \item{name}{The name of the equation.}
  \item{ref}{The equation reference.}
  \item{refShort}{The equation reference shortened.}
  \item{startVal}{The parameters found in the literature with their references.}
  \item{com}{An optional comment about the equation use.}
  \item{id}{An id to identify the equation.}
}
}
\usage{
beta_95
}
\description{
Yin, X., Kropff, M.J., McLaren, G., and Visperas, R.M. (1995) A nonlinear model for crop
  development as a function of temperature. Agricultural and Forest Meteorology 77(1): 1-16.
}
\details{
Equation:
\deqn{rT = e^{mu} * (T - Tb)^{aa} * (Tc - T)^{bb}}{%
      rT = exp(mu) * (T - Tb)^aa * (Tc - T)^bb}

where rT is the development rate, T the temperature, mu, aa,
and bb the model parameters, Tb the base temperature, and Tc the ceiling
temperature.
}
\references{
\doi{10.1016/0168-1923(95)02236-Q}
}
\keyword{datasets}
