% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devRate_utils.R
\name{compDifDays}
\alias{compDifDays}
\title{Compute the inverse of number of days between dates}
\usage{
compDifDays(vecDates, dateFormat = "\%d/\%m/\%y")
}
\arguments{
\item{vecDates}{A vector with dates.}

\item{dateFormat}{The format of dates (see \code{\link[base]{strptime}}).}
}
\value{
A vector with the inverse of the difference between dates.
}
\description{
compDifDays computes the inverse of the difference between dates form a vector made of dates.
}
\examples{
compDifDays(vecDates = c("28/12/15", "12/01/16", "25/01/16", "28/02/16", "15/03/16"))
compDifDays(vecDates = c("28/12/15", "12/01/14", "25/01/16", "28/02/16", "15/03/16"))
compDifDays(vecDates = c("28/12/15", "12/01/16", "25/01/16", "", ""))
}

