% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devRate.R
\docType{package}
\name{devRate}
\alias{devRate}
\alias{devRate-package}
\title{devRate: A package for quantifying the relationships between developmental rate and temperature in ectotherms.}
\description{
devRate: A package for quantifying the relationships between developmental rate and temperature in ectotherms.
}

