% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{devRateEqList}
\alias{devRateEqList}
\title{The list of all available equations of development rate as a function of temperature.}
\format{
An object of class \code{list} of length 37.
}
\usage{
devRateEqList
}
\description{
The list of all available equations of development rate as a function of temperature.
}
\keyword{datasets}
