% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{devRateEqStartVal}
\alias{devRateEqStartVal}
\title{Default starting values for each equation listed in the devRateEqList object.}
\format{
An object of class \code{list} of length 37.
}
\usage{
devRateEqStartVal
}
\description{
Default starting values for each equation listed in the devRateEqList object.
}
\keyword{datasets}
