% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devRate_find.R
\name{devRateFind}
\alias{devRateFind}
\title{Find models for species}
\usage{
devRateFind(orderSP = "", familySP = "", species = "")
}
\arguments{
\item{orderSP}{Find models by Order.}

\item{familySP}{Find models by Family.}

\item{species}{Find models by species (Genus species).}
}
\value{
Nothing.
}
\description{
Find models for species
}
\examples{
devRateFind(orderSP = "Lepidoptera")
devRateFind(familySP = "Gelechiidae")
devRateFind(species = "Tuta absoluta")
}

