% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devRate_info.R
\name{devRateInfo}
\alias{devRateInfo}
\title{Display information about an equation}
\usage{
devRateInfo(eq)
}
\arguments{
\item{eq}{The name of the equation.}
}
\value{
Nothing.
}
\description{
Display information about an equation
}
\examples{
devRateInfo(eq = davidson_44)
devRateInfo(eq = campbell_74)
devRateInfo(eq = taylor_81)
devRateInfo(eq = wang_82)
}
