% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devRate_qualityFits.R
\name{devRateQlBio}
\alias{devRateQlBio}
\title{Biological likelihood of nls fits}
\usage{
devRateQlBio(nlsDR, propThresh = 0.01, eq)
}
\arguments{
\item{nlsDR}{A list of nls objects.}

\item{propThresh}{The proportion of maximal development rate used as a
threshold for estimating CTmin and CTmax when asymptotic equations are used
(default value is 0.01)}

\item{eq}{A list of equations used for nls fitting.}
}
\value{
An object of class \code{data.frame} with development metrics (CTmin,
  Ctmax, Topt) in columns and nls objects in rows.
}
\description{
Return a table of three metrics of development (CTmin, CTmax, Topt)
}
\details{
NULL is returned when nlsDR or df are not a list.
}
\examples{
myDf <- data.frame(T = seq(from = 0, to = 50, by = 10),
                   rT = c(0.001, 0.008, 0.02, 0.03, 0.018, 0.004))
myNLS <- list(devRateModel(eq = damos_08, df = myDf,
                           startValues = list(aa = 1, bb = 1, cc = 1), algo = "LM"),
              devRateModel(eq = kontodimas_04, df = myDf,
                           startValues = list(aa = 1, Tmin = 7, Tmax = 40), algo = "LM"),
              devRateModel(eq = poly2, df = myDf,
                           startValues = list(a0 = 1, a1 = 1, a2 = 1), algo = "LM"))
devRateQlBio(nlsDR = myNLS, propThresh = 0.1, eq = list(damos_08, kontodimas_04, poly2))
}
