% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devRate_harmigera.R
\name{ha_ahmad2024_ls}
\alias{ha_ahmad2024_ls}
\title{Ahmad linear thermal performance curve for the development of
  Helicoverpa armigera}
\usage{
ha_ahmad2024_ls(plotfig = TRUE)
}
\arguments{
\item{plotfig}{A Boolean used to return the experimental points and the
equation fitted in the article.}
}
\value{
A list with the equation used, and a list of model parameters for
  the different life stages considered in the article.
}
\description{
Linear development performance curve from four
  experimental temperatures (14, 16, 18, 20, 22, 25, 27, 30, 32, 35 and 36
  degrees Celsius).
}
\details{
This work is part of the ACOMPLI project. The ACOMPLI project
  is part of the Strategic Action Plan for the anticipation of the potential
  European withdrawal of active substances and the development of
  alternative crop protection techniques (PARSADA). It is financed by
  ecological planning funds. The French Ministry of Agriculture cannot be
  held responsible for the content of this package.
}
\examples{
  mymodel <- ha_ahmad2024_ls(plotfig = FALSE)
}
\seealso{
Ahmad S. (2024) Temperature dependent survivorship and development
  of Helicoverpa armigera (Hubner) (Lepidoptera-Noctuidae) on chickpea.
  Preprint. https://doi.org/10.21203/rs.3.rs-4845823/v1
}
