% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devRate_harmigera.R
\name{ha_bartekova2006}
\alias{ha_bartekova2006}
\title{Bartekova and Praslicka linear thermal performance curve for the
  developmentof Helicoverpa armigera}
\usage{
ha_bartekova2006(plotfig = TRUE)
}
\arguments{
\item{plotfig}{A Boolean used to return the experimental points and the
equation fitted in the article.}
}
\value{
A list with the equation used, and a list of model parameters for
  the different life stages considered in the article.
}
\description{
Linear development performance curve for eggs, larvae and
  pupae from three experimental temperatures (20, 25, and 30 degrees
  Celsius).
}
\details{
This work is part of the ACOMPLI project. The ACOMPLI project
  is part of the Strategic Action Plan for the anticipation of the potential
  European withdrawal of active substances and the development of
  alternative crop protection techniques (PARSADA). It is financed by
  ecological planning funds. The French Ministry of Agriculture cannot be
  held responsible for the content of this package.
}
\examples{
  mymodel <- ha_bartekova2006(plotfig = FALSE)
}
\seealso{
Bartekova, A., and Praslicka, J. (2006). The effect of ambient
  temperature on the development of cotton bollworm (Helicoverpa armigera
  Hubner, 1808). Plant Protection Science, 42(4), 135.
  https://doi.org/10.17221/2768-PPS
}
