% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devRate_harmigera.R
\name{ha_foley1981}
\alias{ha_foley1981}
\title{Foley linear thermal performance curve for the post-diapausing and
  non-diapausing pupae development of Helicoverpa armigera}
\usage{
ha_foley1981(plotfig = TRUE)
}
\arguments{
\item{plotfig}{A Boolean used to return the experimental points and the
equation fitted in the article.}
}
\value{
A list with the equation used, and a list of model parameters for
  the different life stages considered in the article.
}
\description{
Linear development performance curve for post-diapausing and
  non-diapausing pupae from 3 to 4 experimental temperatures (20, 24,
  28, and 32 degrees Celsius).
}
\details{
This work is part of the ACOMPLI project. The ACOMPLI project
  is part of the Strategic Action Plan for the anticipation of the potential
  European withdrawal of active substances and the development of
  alternative crop protection techniques (PARSADA). It is financed by
  ecological planning funds. The French Ministry of Agriculture cannot be
  held responsible for the content of this package.
}
\examples{
  mymodel <- ha_foley1981(plotfig = FALSE)
}
\seealso{
Foley, D. H. (1981). Pupal development rate of Heliothis armiger
  (Hubner)(Lepidoptera: Noctuidae) under constant and alternating
  temperatures. Australian Journal of Entomology, 20(1), 13-20.
  https://doi.org/10.1111/j.1440-6055.1981.tb00993.x
}
