% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devRate_harmigera.R
\name{ha_kay1981_ls}
\alias{ha_kay1981_ls}
\title{Kay linear thermal performance curve for the egg development of
  Helicoverpa armigera}
\usage{
ha_kay1981_ls(plotfig = TRUE)
}
\arguments{
\item{plotfig}{A Boolean used to return the experimental points and the
equation fitted in the article.}
}
\value{
A list with the equation used, and a list of model parameters for
  the different life stages considered in the article.
}
\description{
Linear development performance curve for eggs from ten
  experimental temperatures (8, 10, 13.3, 17.8, 20.8, 24.4, 27.2, 31.4,
  35, 39.4 degrees Celsius).
}
\details{
This work is part of the ACOMPLI project. The ACOMPLI project
  is part of the Strategic Action Plan for the anticipation of the potential
  European withdrawal of active substances and the development of
  alternative crop protection techniques (PARSADA). It is financed by
  ecological planning funds. The French Ministry of Agriculture cannot be
  held responsible for the content of this package.
}
\examples{
  mymodel <- ha_kay1981_ls(plotfig = FALSE)
}
\seealso{
Kay, I. R. (1981). The effect of constant temperatures on the
  development time of eggs of Heliothis armiger (Hubner) (Lepidoptera:
  Noctuidae). Australian Journal of Entomology, 20(2), 155-156.
  https://doi.org/10.1111/j.1440-6055.1981.tb01020.x
}
