% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devRate_harmigera.R
\name{ha_mironidis2008_ls}
\alias{ha_mironidis2008_ls}
\title{Mironidis and Savopoulou-Soultani linear thermal performance curve for the
  development of Helicoverpa armigera}
\usage{
ha_mironidis2008_ls(plotfig = TRUE)
}
\arguments{
\item{plotfig}{A Boolean used to return the experimental points and the
equation fitted in the article.}
}
\value{
A list with the equation used, and a list of model parameters for
  the different life stages considered in the article.
}
\description{
Linear development performance curve for eggs, larvae and
  pupae from twelve experimental temperatures (10, 12.5, 15, 17.5, 20,
  25, 27.5, 30, 32.5, 35, 37.5, and 40 degrees Celsius).
}
\details{
This work is part of the ACOMPLI project. The ACOMPLI project
  is part of the Strategic Action Plan for the anticipation of the potential
  European withdrawal of active substances and the development of
  alternative crop protection techniques (PARSADA). It is financed by
  ecological planning funds. The French Ministry of Agriculture cannot be
  held responsible for the content of this package.
}
\examples{
  mymodel <- ha_mironidis2008_ls(plotfig = FALSE)
}
\seealso{
Mironidis, G. K., and Savopoulou-Soultani, M. (2008). Development,
  survivorship, and reproduction of Helicoverpa armigera (Lepidoptera:
  Noctuidae) under constant and alternating temperatures. Environmental
  Entomology, 37(1), 16-28. https://doi.org/10.1093/ee/37.1.16
}
