% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devRate_harmigera.R
\name{ha_noorulane2018_nls}
\alias{ha_noorulane2018_nls}
\title{Noor-ul-Ane et al. non-linear thermal performance curve for the development
  of Helicoverpa armigera}
\usage{
ha_noorulane2018_nls(plotfig = TRUE)
}
\arguments{
\item{plotfig}{A Boolean used to return the experimental points and the
equation fitted in the article.}
}
\value{
A list with the equation used, and a list of model parameters for
  the different life stages considered in the article.
}
\description{
Non-linear Briere2 development performance curve from ten
  experimental temperatures (10, 15, 17.5, 20, 25, 27.5, 30, 35, 37.5
  and 40 degrees Celsius). Experimental development data were retrieve from
  Figure 1 using plotdigitizer.com.
}
\details{
This work is part of the ACOMPLI project. The ACOMPLI project
  is part of the Strategic Action Plan for the anticipation of the potential
  European withdrawal of active substances and the development of
  alternative crop protection techniques (PARSADA). It is financed by
  ecological planning funds. The French Ministry of Agriculture cannot be
  held responsible for the content of this package.
}
\examples{
  mymodel <- ha_noorulane2018_nls(plotfig = FALSE)
}
\seealso{
Noor-ul-Ane M., Mirhosseini M. A., Crickmore N., Saeed S.,
  Noor I., Zalucki M. P. (2018). Temperature-dependent development of
  Helicoverpa armigera (Hubner) (Lepidoptera: Noctuidae) and its larval
  parasitoid, Habrobracon hebetor (Say) (Hymenoptera: Braconidae):
  implications for species interactions. Bulletin of Entomological Research
  108, 295–304. https://doi.org/10.1017/S0007485317000724
}
