% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devRate_harmigera.R
\name{ha_qureshi1999_ls}
\alias{ha_qureshi1999_ls}
\title{Qureshi et al. linear thermal performance curve for the development of
  Helicoverpa armigera}
\usage{
ha_qureshi1999_ls(plotfig = TRUE)
}
\arguments{
\item{plotfig}{A Boolean used to return the experimental points and the
equation fitted in the article.}
}
\value{
A list with the equation used, and a list of model parameters for
  the different life stages considered in the article.
}
\description{
Linear development performance curve from four
  experimental temperatures (15, 20, 25,
  and 30 degrees Celsius).
}
\details{
This work is part of the ACOMPLI project. The ACOMPLI project
  is part of the Strategic Action Plan for the anticipation of the potential
  European withdrawal of active substances and the development of
  alternative crop protection techniques (PARSADA). It is financed by
  ecological planning funds. The French Ministry of Agriculture cannot be
  held responsible for the content of this package.
}
\examples{
  mymodel <- ha_qureshi1999_ls(plotfig = FALSE)
}
\seealso{
Qureshi, M. H., T. Murai, H. Yoshida, T. Shiraga and H. Tsumuki
  (1999) Effects of photoperiod and temperature on development and diapause
  induction in the Okayama population of Helicoverpa armigera (Hb)
  (Lepidoptera: Noctuidae). Appl. Entomol. Zool. 34: 327–331.
  https://doi.org/10.1303/aez.34.327
}
