% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hansen_11}
\alias{hansen_11}
\title{Hansen equation of development rate as a function of temperature.}
\format{A list of eight elements describing the equation.
\describe{
  \item{eq}{The equation as a formula object.}
  \item{eqAlt}{The equation as a string.}
  \item{name}{The equation name.}
  \item{ref}{The equation reference.}
  \item{refShort}{The equation reference shortened.}
  \item{startVal}{The parameter values found in the literature with their references.}
  \item{com}{An otional comment about the equation use.}
  \item{id}{An id to identificate the equation.}
}}
\source{
\url{http://dx.doi.org/10.1016/j.jinsphys.2011.06.011}
}
\usage{
hansen_11
}
\description{
Hansen, E.M., Bentz, B.J., Powell, J.A., Gray, D.R., and Vandygriff, J.C. (2011) Prepupal
diapause and instar IV developmental rates of the spruce beetle, Dendroctonus rufipennis
(Coleoptera: Curculionidae, Scolytinae). Journal of insect physiology 57(10): 1347-57.
}
\details{
Equation:
rT ~ p2 * ( (exp(p3 * (T - p1)) - 1) * (exp(p3 * (p5 - p1)) - 1) * exp((T - p5) / p4) )

where rT is the development rate, T the temperature, p1 the lower developmental
treshold, p2 the peak rate control parameter, p3 the low temperature acceleration of
rates, p4 the width of upper thermal boundary layer, and p5 the upper developmental
threshold.
}
\keyword{datasets}

