% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kontodimas_04}
\alias{kontodimas_04}
\title{Kontodimas et al. equation of development rate as a function of temperature.}
\format{
A list of eight elements describing the equation.
\describe{
  \item{eq}{The equation (formula object).}
  \item{eqAlt}{The equation (string).}
  \item{name}{The name of the equation.}
  \item{ref}{The equation reference.}
  \item{refShort}{The equation reference shortened.}
  \item{startVal}{The parameters found in the literature with their references.}
  \item{com}{An optional comment about the equation use.}
  \item{id}{An id to identify the equation.}
}
}
\usage{
kontodimas_04
}
\description{
Kontodimas, D.C., Eliopoulos, P.A., Stathas, G.J. and Economou, L.P. (2004) Comparative
temperature-dependent development of Nephus includens (Kirsch) and Nephus bisignatus
(Boheman)(Coleoptera: Coccinellidae) preying on Planococcus citri
(Risso)(Homoptera: Pseudococcidae): evaluation of a linear and various nonlinear models
using specific criteria. Environmental Entomology 33(1): 1-11.
}
\details{
Equation:
\deqn{rT = aa * (T - Tmin)^{2} * (Tmax - T)}{%
      rT = aa * (T - Tmin)^2 * (Tmax - T)}

where rT is the development rate, T the temperature, Tmin the minimum
temperature, Tmax the maximum temperature, and aa a constant.
}
\references{
\doi{10.1603/0046-225X-33.1.1}
}
\keyword{datasets}
