% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{schoolfield_81}
\alias{schoolfield_81}
\title{Schoolfield et al. equation of development rate as a function of temperature.}
\format{
A list of eight elements describing the equation.
\describe{
  \item{eq}{The equation (formula object).}
  \item{eqAlt}{The equation (string).}
  \item{name}{The name of the equation.}
  \item{ref}{The equation reference.}
  \item{refShort}{The equation reference shortened.}
  \item{startVal}{The parameters found in the literature with their references.}
  \item{com}{An optional comment about the equation use.}
  \item{id}{An id to identify the equation.}
}
}
\usage{
schoolfield_81
}
\description{
Schoolfield, R., Sharpe, P. & Magnuson, C. (1981) Non-linear regression of biological
temperature-dependent rate models based on absolute reaction-rate theory.
Journal of theoretical biology, 88, 719-731.
}
\details{
Equation:
\deqn{rT = \frac{p25 * \frac{T + 273.16}{298} * e^{\frac{aa}{1.987} * (\frac{1}{298} - \frac{1}{T + 273.16})}}{1 + e^{\frac{bb}{1.987} * (\frac{1}{cc} - \frac{1}{T + 273.16})} + e^{\frac{dd}{1.987} * (\frac{1}{ee} - \frac{1}{T + 273.16})}}}{%
      rT = (p25 * (T + 273.16)/298 * exp(aa/1.987 * (1/298 - 1/(T + 273.16)))) / (1 + exp(bb/1.987 * (1/cc - 1/(T + 273.16))) + exp(dd/1.987 * (1/ee - 1/(T + 273.16))))}

where rT is the development rate, T the temperature, p25 the development
rate at 25 degree Celsius assuming no enzyme inactivation, aa the enthalpy of
activation of the reaction that is catalyzed by the enzyme, bb the change in
enthalpy associated with low temperature inactivation of the enzyme, cc the
temperature at which the enzyme is 1/2 active and 1/2 low temperature inactive,
dd the change in enthalpy associated with high temperature inactivation of the enzyme,
and ee the temperature at which the enzyme is 1/2 active and 1/2 high temperature
inactive.
}
\references{
\doi{10.1016/0022-5193(81)90246-0}
}
\keyword{datasets}
