% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wang_82}
\alias{wang_82}
\title{Wang et al. equation of development rate as a function of temperature.}
\format{A list of eight elements describing the equation.
\describe{
  \item{eq}{The equation as a formula object.}
  \item{eqAlt}{The equation as a string.}
  \item{name}{The equation name.}
  \item{ref}{The equation reference.}
  \item{refShort}{The equation reference shortened.}
  \item{startVal}{The parameter values found in the literature with their references.}
  \item{com}{An otional comment about the equation use.}
  \item{id}{An id to identificate the equation.}
}}
\source{
\url{http://en.cnki.com.cn}
}
\usage{
wang_82
}
\description{
Wang, R., Lan, Z. and Ding, Y. (1982) Studies on mathematical models of the relationship
between insect development and temperature. Acta Ecol. Sin, 2, 47-57.
}
\details{
Equation:
rT ~
}
\keyword{datasets}

