% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{analytis_77}
\alias{analytis_77}
\title{Analytis equation of development rate as a function of temperature.}
\format{A list of eight elements describing the equation.
\describe{
  \item{eq}{The equation as a formula object.}
  \item{eqAlt}{The equation as a string.}
  \item{name}{The equation name.}
  \item{ref}{The equation reference.}
  \item{refShort}{The equation reference shortened.}
  \item{startVal}{The parameter values found in the literature with their references.}
  \item{com}{An otional comment about the equation use.}
  \item{id}{An id to identificate the equation.}
}}
\source{
\url{http://dx.doi.org/10.1111/j.1439-0434.1977.tb02886.x}
}
\usage{
analytis_77
}
\description{
Analytis, S. (1977) Uber die Relation zwischen biologischer Entwicklung und Temperatur bei
phytopathogenen Pilzen. Journal of Phytopathology 90(1): 64-76.
}
\details{
Equation:
rT ~
}
\keyword{datasets}

