% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{briere1_99}
\alias{briere1_99}
\title{Briere et al equation 1 of development rate as a function of temperature.}
\format{
A list of eight elements describing the equation.
\describe{
  \item{eq}{The equation (formula object).}
  \item{eqAlt}{The equation (string).}
  \item{name}{The name of the equation.}
  \item{ref}{The equation reference.}
  \item{refShort}{The equation reference shortened.}
  \item{startVal}{The parameters found in the literature with their references.}
  \item{com}{An optional comment about the equation use.}
  \item{id}{An id to identify the equation.}
}
}
\usage{
briere1_99
}
\description{
Briere, J.F., Pracros, P., le Roux, A.Y. and Pierre, S. (1999) A novel rate model of
temperature-dependent development for arthropods. Environmental Entomology, 28, 22-29.
}
\details{
Equation:
\deqn{rT = aa * T * (T - Tmin) * (Tmax - T)^{\frac{1}{2}}}{%
      rT = aa * T * (T - Tmin) * (Tmax - T)^(1 / 2)}

where rT is the development rate, T the temperature, Tmin the low
temperature developmental threshold, Tmax the lethal temperature, and aa
an empirical constant.
}
\references{
\doi{10.1093/ee/28.1.22}
}
\keyword{datasets}
