% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devRate.R
\docType{package}
\name{devRate}
\alias{devRate}
\alias{devRate-package}
\title{devRate: A package for quantifying the relationship between developmental rate and temperature in ectotherms.}
\description{
The devRate package allows quantifying the relationships between developmental rate and temperature in ectotherm organisms.
}
\details{
### How to use the package

You can use the package:

- to get developmental rate curves as a function of temperature for a specific organism (hundred of examples from the literature are included in the package);

- to know wich equations exists and which are most used in the literature; and

- to relate developmental rate with temperature from your empirical data, using the equations from the package.

### Installation instructions

install.packages("devRate")

### Overview

The devRate package provides three categories of functions:

- to find developmental rate information about a specific organism (Order, Family, Genus, species): which equations were used and which are the asociated parameters (e.g., helpful to estimate starting values for your empirical data sets);

- to relate developmental rate and temperature; and

- to plot your empirical data sets and the associated fitted model, and/or to plot develpomental curves from the literature.

Author's affiliation: IRD, UMR EGCE, Univ.ParisSud-CNRS-IRD-Univ.ParisSaclay
}
