% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{devRateEqStartVal}
\alias{devRateEqStartVal}
\title{The list of start values necessary for the fitting of every equations.}
\format{
An object of class \code{list} of length 37.
}
\usage{
devRateEqStartVal
}
\description{
The list of start values necessary for the fitting of every equations.
}
\keyword{datasets}
