% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devRate_nls.R
\name{devRateModel}
\alias{devRateModel}
\title{Compute non-linear regression}
\usage{
devRateModel(eq, temp, devRate, startValues, ...)
}
\arguments{
\item{eq}{The name of the equation.}

\item{temp}{The temperature.}

\item{devRate}{The developmental rate \code{(days)^-1}}

\item{startValues}{Starting values for the regression.}

\item{...}{Additional arguments for the nls function.}
}
\description{
Compute non-linear regression
}

