% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{harcourtYee_82}
\alias{harcourtYee_82}
\title{Harcourt and Yee equation of developmental rate as a function of temperature.}
\format{A list of eight elements describing the equation.
\describe{
  \item{eq}{The equation as a formula object.}
  \item{eqAlt}{The equation as a string.}
  \item{name}{The equation name.}
  \item{ref}{The equation reference.}
  \item{refShort}{The equation reference shortened.}
  \item{startVal}{The parameter values found in the literature with their references.}
  \item{com}{An otional comment about the equation use.}
  \item{id}{An id to identificate the equation.}
}}
\source{
\url{http://dx.doi.org/10.1093/ee/11.3.581}
}
\usage{
harcourtYee_82
}
\description{
Harcourt, D. and Yee, J. (1982) Polynomial algorithm for predicting the duration of insect
life stages. Environmental Entomology, 11, 581-584.
}
\keyword{datasets}

