% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taylor_81}
\alias{taylor_81}
\title{Taylor equation of developmental rate as a function of temperature.}
\format{A list of eight elements describing the equation.
\describe{
  \item{eq}{The equation as a formula object.}
  \item{eqAlt}{The equation as a string.}
  \item{name}{The equation name.}
  \item{ref}{The equation reference.}
  \item{refShort}{The equation reference shortened.}
  \item{startVal}{The parameter values found in the literature with their references.}
  \item{com}{An otional comment about the equation use.}
  \item{id}{An id to identificate the equation.}
}}
\source{
\url{http://www.jstor.org/stable/2460694}
}
\usage{
taylor_81
}
\description{
Taylor, F. (1981) Ecology and evolution of physiological time in insects.
American Naturalist, 1-23.
Lamb, RJ. (1992) Developmental rate of Acyrthosiphon pisum (Homoptera: Aphididae) at low
temperatures: implications for estimating rate parameters for insects.
Environmental Entomology 21(1): 10-19.
}
\keyword{datasets}

